Refactor tauri related stuff into RuntimeProvider
