+++ studio/src/main.tsx
@@ -1,7 +1,7 @@
+import { RuntimeProvider } from "@/components/RuntimeProvider";
 import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App.tsx";
-import { RuntimeProvider } from "./tauri";
 import "./index.css";
 
 const root = document.getElementById("root");
