--- studio/src/tauri/index.ts
@@ -1,1 +1,0 @@
-export * from "./RuntimeProvider/RuntimeProvider";
